requirements.txt
pyserial==3.5
PySide6==6.6.1
pandas==2.1.4
matplotlib==3.8.2
openpyxl==3.1.2
jinja2==3.1.3
